// assets/scripts/types/Enum.ts
export enum GameMode {
    ADVENTURE = 'adventure',
    SURVIVAL = 'survival',
    MINI_GAME = 'mini_game',
    ZEN_GARDEN = 'zen_garden'
}

export enum SceneType {
    MENU = 'menu',
    ADVENTURE_1_DAY = 'adventure_1_day',
    ADVENTURE_2_NIGHT = 'adventure_2_night',
    ADVENTURE_3_POOL = 'adventure_3_pool',
    ADVENTURE_4_ROOF = 'adventure_4_roof',
    ADVENTURE_5_FOG = 'adventure_5_fog',
    SURVIVAL_NORMAL = 'survival_normal',
    SURVIVAL_ENDLESS = 'survival_endless'
}

export enum PlantType {
    SUNFLOWER = 'sunflower',
    PEASHOOTER = 'peashooter',
    SNOW_PEA = 'snow_pea',
    WALL_NUT = 'wall_nut',
    CHERRY_BOMB = 'cherry_bomb',
    REPEATER = 'repeater',
    CHOMPER = 'chomper',
    TORCHWOOD = 'torchwood',
    ICEBERG_LETTUCE = 'iceberg_lettuce',
    SQUASH = 'squash',
    THREEPEATER = 'threepeater',
    SPLIT_PEA = 'split_pea',
    JALAPENO = 'jalapeno',
    SPIKE_WEED = 'spike_weed',
    TALL_NUT = 'tall_nut',
    TWIN_SUNFLOWER = 'twin_sunflower',
    GLOOM_SHROOM = 'gloom_shroom',
    CATTAIL = 'cattail',
    WINTER_MELON = 'winter_melon',
    GATLING_PEA = 'gatling_pea',
    PUFF_SHROOM = 'puff_shroom',
    SUN_SHROOM = 'sun_shroom',
    FUME_SHROOM = 'fume_shroom',
    GRAVE_BUSTER = 'grave_buster',
    HYPNO_SHROOM = 'hypno_shroom',
    SCAREDY_SHROOM = 'scaredy_shroom',
    ICE_SHROOM = 'ice_shroom',
    DOOM_SHROOM = 'doom_shroom',
    LILY_PAD = 'lily_pad',
    SPIKE_ROCK = 'spike_rock',
    MAGNET_SHROOM = 'magnet_shroom',
    CABBAGE_PULT = 'cabbage_pult',
    FLOWER_POT = 'flower_pot',
    KERNEL_PULT = 'kernel_pult',
    COFFEE_BEAN = 'coffee_bean',
    GARLIC = 'garlic',
    UMBRELLA_LEAF = 'umbrella_leaf',
    MARIGOLD = 'marigold',
    MELON_PULT = 'melon_pult',
    COB_CANNON = 'cob_cannon',
    CACTUS = 'cactus',
    BLOVER = 'blover',
    SPLASH_SHROOM = 'splash_shroom',
    SQUIRREL = 'squirrel'
}

export enum PlantCategory {
    PRODUCER = 'producer',
    ATTACKER = 'attacker',
    DEFENDER = 'defender',
    SPECIAL = 'special'
}

export enum ZombieType {
    NORMAL = 'normal',
    FLAG = 'flag',
    CONE = 'cone',
    BUCKET = 'bucket',
    POLE_VAULTING = 'pole_vaulting',
    NEWSPAPER = 'newspaper',
    SCREEN_DOOR = 'screen_door',
    FOOTBALL = 'football',
    DANCING = 'dancing',
    BACKUP_DANCER = 'backup_dancer',
    DUCKY_TUBE = 'ducky_tube',
    SNORKEL = 'snorkel',
    ZOMBONI = 'zomboni',
    JACK_IN_THE_BOX = 'jack_in_the_box',
    BALLOON = 'balloon',
    DIGGER = 'digger',
    POGO = 'pogo',
    BUNGEE = 'bungee',
    LADDER = 'ladder',
    CATAPULT = 'catapult',
    GARGANTUAR = 'gargantuar',
    IMP = 'imp',
    DR_ZOMBOSS = 'dr_zomboss'
}

export enum ProjectileType {
    NORMAL_PEA = 'normal_pea',
    SNOW_PEA = 'snow_pea',
    FIRE_PEA = 'fire_pea',
    MELON = 'melon',
    WINTER_MELON = 'winter_melon',
    COB = 'cob'
}

export enum SunType {
    NATURAL = 'natural',
    PRODUCED = 'produced'
}

export enum GameState {
    MENU = 'menu',
    PLAYING = 'playing',
    PAUSED = 'paused',
    VICTORY = 'victory',
    DEFEAT = 'defeat'
}

export enum DamageType {
    NORMAL = 'normal',
    ICE = 'ice',
    FIRE = 'fire',
    INSTANT = 'instant'
}